backend/requirements.txt
# Backend application requirements
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
sse-starlette==2.2.1

# GitHub Copilot SDK
github-copilot-sdk>=0.1.30

# Azure Identity & Services
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-mgmt-resource==23.2.0
azure-mgmt-network==28.0.0
azure-mgmt-storage==21.2.0
azure-mgmt-sql==4.0.0
azure-mgmt-monitor==6.0.0
azure-mgmt-security==7.0.0
msgraph-sdk==1.16.0

# OpenTelemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
azure-monitor-opentelemetry-exporter==1.0.0b33
opentelemetry-exporter-otlp==1.29.0

# Utilities
structlog==24.4.0
jinja2==3.1.5
markdown==3.7
weasyprint==63.1
aiofiles==24.1.0
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
respx==0.22.0
